remove userProfile folder
